Bugfix/trac10510.
@VemundH  ; please remind me tomorrow to merge to master (assuming the tests cycle successfully) - and update the Trac report.

+++ include/boost/uuid/detail/uuid_x86.hpp
@@ -100,7 +100,7 @@ inline bool operator< (uuid const& lhs, uuid const& rhs) BOOST_NOEXCEPT
     cmp = (cmp - 1u) ^ cmp;
     rcmp = (rcmp - 1u) ^ rcmp;
 
-    return static_cast< uint16_t >(cmp) < static_cast< uint16_t >(rcmp);
+    return cmp < rcmp;
 }
 
 } // namespace uuids
